// lib/types.d.ts
export interface components {
  schemas: {
// ...
    Contract: {
      /** @description ID of the contract. */
      id: string;
      /** @description The symbol of the faction that this contract is for. */
      factionSymbol: string;
      /**
       * @description Type of contract.
       * @enum {string}
       */
      type: "PROCUREMENT" | "TRANSPORT" | "SHUTTLE";
      terms: components["schemas"]["ContractTerms"];
      /**
       * @description Whether the contract has been accepted by the agent
       * @default false
       */
      accepted: boolean;
      /**
       * @description Whether the contract has been fulfilled
       * @default false
       */
      fulfilled: boolean;
      /**
       * Format: date-time
       * @deprecated
       * @description Deprecated in favor of deadlineToAccept
       */
      expiration: string;
      /**
       * Format: date-time
       * @description The time at which the contract is no longer available to be accepted
       */
      deadlineToAccept?: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: {
  };
  headers: never;
  pathItems: never;
}